assignment3/pyproject.toml
[build-system]
requires = [
    "setuptools",
    # 4110 only:
    # "cython",
    # "numpy==1.21.*"
]
build-backend = "setuptools.build_meta"

[project]
version = "0.1.0"
requires-python = ">=3.7"
license = {text = "MIT License"}
name = "instapy"
description = "filter images"
readme = "README.md"
dependencies = [ 'numpy == 1.21.5',
    'pillow == 9.0.0' ,
    'numba == 0.55.1' ,
    'line-profiler == 3.5.1' ,
]


[project.scripts]
# instapy = "instapy.cli:main"
